Convert Simplifiers to Plugins
This allows us to extensibly discover simplifiers
