+++ scijava/scijava-ops/src/main/java/org/scijava/ops/simplify/Simplifier.java
@@ -3,16 +3,17 @@ package org.scijava.ops.simplify;
 
 import java.lang.reflect.Type;
 
+import org.scijava.plugin.SciJavaPlugin;
 import org.scijava.types.Types;
 
 /**
  * An object that can convert between a general and precise type.
  * 
- * @author G
+ * @author Gabriel Selzer
  * @param <G> - the general type
  * @param <P> - the precise type
  */
-public interface Simplifier<G, P> {
+public interface Simplifier<G, P> extends SciJavaPlugin {
 
 	public G simplify(P p);
